+++ api/service/syncing/downloader/gen.sh
