+++ node/node.go
@@ -24,10 +24,10 @@ import (
 	clientService "github.com/harmony-one/harmony/api/client/service"
 	proto_discovery "github.com/harmony-one/harmony/api/proto/discovery"
 	proto_node "github.com/harmony-one/harmony/api/proto/node"
-	"github.com/harmony-one/harmony/api/services/explorer"
-	"github.com/harmony-one/harmony/api/services/syncing"
-	"github.com/harmony-one/harmony/api/services/syncing/downloader"
-	downloader_pb "github.com/harmony-one/harmony/api/services/syncing/downloader/proto"
+	"github.com/harmony-one/harmony/api/service/explorer"
+	"github.com/harmony-one/harmony/api/service/syncing"
+	"github.com/harmony-one/harmony/api/service/syncing/downloader"
+	downloader_pb "github.com/harmony-one/harmony/api/service/syncing/downloader/proto"
 	bft "github.com/harmony-one/harmony/consensus"
 	"github.com/harmony-one/harmony/core"
 	"github.com/harmony-one/harmony/core/types"
